geo_fencing/requirements.txt
psycopg[binary]
shapely
geopy
python-dotenv
pandas
geopandas